test: move `output()` test to `test/output.js`
> #1
